split the test_ext.py to a separate testing modules
